requirements.txt
fastapi[all]
fastapi-mail
tortoise-orm
mysql-connector-python
aiomysql
python-dotenv
python-jose[cryptography]
aiofiles
requests